requirements.txt
transformers>=4.38,<5.0
torch>=2.1,<3.0
flask>=2.2,<3.0
requests>=2.25,<3.0
python-dotenv>=1.0,<2.0
beautifulsoup4>=4.12,<5.0
python-telegram-bot==13.15
